Create a fact model
This will be the core component of this application. It will store
facts that are small, regularly changing, pieces of information that
are used in multiple places across GOV.UK.

The basic set of tests just makes sure we can create a Fact object
with the respective default values and that it gets persisted.